requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
tensorflow-cpu==2.15.0
pillow==10.0.1
numpy<2.0.0